+++ persistent-sqlite/test/main.hs
@@ -22,6 +22,7 @@ import qualified EmptyEntityTest
 import qualified EmbedOrderTest
 import qualified EmbedTest
 import qualified EquivalentTypeTest
+import qualified ForeignKey
 import qualified HtmlTest
 import qualified LargeNumberTest
 import qualified MaxLenTest
@@ -141,6 +142,7 @@ main = do
       , MaxLenTest.maxlenMigrate
       , Recursive.recursiveMigrate
       , CompositeTest.compositeMigrate
+      , ForeignKey.compositeMigrate
       , MigrationTest.migrationMigrate
       , PersistUniqueTest.migration
       , RenameTest.migration
@@ -205,16 +207,17 @@ main = do
     CustomPrimaryKeyReferenceTest.specsWith db
     MigrationColumnLengthTest.specsWith db
     EquivalentTypeTest.specsWith db
+    ForeignKey.specsWith db
     TransactionLevelTest.specsWith db
     MigrationTest.specsWith db
 
     it "issue #328" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        runMigration migrateAll
+        runMigrationSilent migrateAll
         _ <- insert . Test $ read "2014-11-30 05:15:25.123Z"
         [Single x] <- rawSql "select strftime('%s%f',time) from test" []
         liftIO $ x `shouldBe` Just ("141732452525.123" :: String)
     it "issue #339" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        runMigration migrateAll
+        runMigrationSilent migrateAll
         now <- liftIO getCurrentTime
         tid <- insert $ Test now
         Just (Test now') <- get tid
@@ -225,15 +228,16 @@ main = do
         Sqlite.close conn
         return ()
     it "issue #527" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        runMigration migrateAll
+        runMigrationSilent migrateAll
         insertMany_ $ replicate 1000 (Test $ read "2014-11-30 05:15:25.123Z")
 
     it "properly migrates to a composite primary key (issue #669)" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        runMigration compositeSetup
-        runMigration compositeMigrateTest
+        runMigrationSilent compositeSetup
+        runMigrationSilent compositeMigrateTest
+        pure ()
 
     it "afterException" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        runMigration testMigrate
+        runMigrationSilent testMigrate
         let catcher :: forall m. Monad m => SomeException -> m ()
             catcher _ = return ()
         _ <- insert $ Person "A" 0 Nothing